Get __version__ directly from _version in setup.py
Importing `__version__` from the `eventbrite` module causes the `client` module to be loaded, which in turn tries to import `requests`. If `requests` is not installed, we get an `ImportError`.

Although the `setup.py` lists `requests` as a requirement, it falls over before it can be installed.

This changes the import to come from `eventbrite._version`, thus sidestepping the issue.